src/requirements.txt
langchain_core==0.3.40
langchain_gigachat==0.3.4
python-dotenv==1.0.1
streamlit==1.42.2